Move from Rc to arena (OperandContext) allocated operands
This reduced full analysis time of samase_scarf by about 20% - and
not all of that time is related to the execution simulation.

..Pretty much all of the operand-related api breaks slightly.
All instances of `Rc<Operand>` and `&Rc<Operand>` need to be converted to
`Operand<'e>` which is just a pointer with free copies.

Equality checks for Operands are now also just a pointer comparision
(Not much improvement as before they were pointer comparision for proving
equality and cached hash comparision to prove inequality)

Deserialization unfortunately is a bit more cumbersome and required writing
out a bunch of the code to get operands allocated on OperandContext.