The better merge caching was poorly thought, out remove it for now
And add a test for a bug in it.

At least the other merging improvements have already helped the worst
cases a bit, though this (bugged) advanced cached helped a lot more.
There seemed to be a possibility about 60-70% time cut in analysis merging
edge cases if the idea can be improved without a big performance hit.