Skip exec state merges when only single branch leads to current branch
That is, assume `merge(old, new)` can be replaced with just `new` when
this is not a merge of two separate execution paths joining.
This helps performance a lot, and effectively behaves the same as before.

But this will likely add some extra branches to be analyzed, e.g. if branch
1 gets reanalyzed due to only `eax` being undefined, but then it zeroes
`eax`, so state doesn't change for any following branches. Before the
following branches would not have been reanalyzed, now they will be if
the only branch leading to them is branch 1.

Since this helps performance so much, should probably try to move even
more towards Analyzer which does as few full merges as possible, and
assumes need for branches to be rechecked depend on if there are new
predecessors in the cfg?

Also removes quite pointless Debug derive from Cfg.