Don't ever realize carry flag in add_resolved_constraint (for now)
add_resolved_constraint only does the flag updates for == constraint,
which almost never is useful for carry anyway. If in future doing that for
> constraint is useful, the carry can be added for that. Only regresses
edge cases such as carry for `x - 1` or `x - u64::MAX` which get
simplified to equality since only one value of `x` would set/unset carry.